Fix: Adjusted page structure as requested
